Move problems and tests into a new package
