+++ src/main/java/net/projecteuler/problems/Problem001.java
@@ -1,4 +1,4 @@
-package net.projecteuler;
+package net.projecteuler.problems;
 
 import java.util.stream.IntStream;
