+++ src/main/java/net/projecteuler/problems/Problem002.java
@@ -1,4 +1,4 @@
-package net.projecteuler;
+package net.projecteuler.problems;
 
 import java.util.HashSet;
 import java.util.Set;
