+++ src/main/java/net/projecteuler/problems/Problem006.java
@@ -1,4 +1,4 @@
-package net.projecteuler;
+package net.projecteuler.problems;
 
 public class Problem006 {
