+++ src/main/java/net/projecteuler/problems/Problem007.java
@@ -1,4 +1,4 @@
-package net.projecteuler;
+package net.projecteuler.problems;
 
 import java.util.ArrayList;
 import java.util.List;
