+++ src/test/java/net/projecteuler/problems/Problem002Test.java
@@ -1,4 +1,4 @@
-package net.projecteuler;
+package net.projecteuler.problems;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
