+++ src/test/java/net/projecteuler/problems/Problem003Test.java
@@ -1,4 +1,4 @@
-package net.projecteuler;
+package net.projecteuler.problems;
 
 import java.util.Collections;
 import org.junit.jupiter.api.Assertions;
