+++ src/test/java/net/projecteuler/problems/Problem005Test.java
@@ -1,4 +1,4 @@
-package net.projecteuler;
+package net.projecteuler.problems;
 
 import java.util.function.Supplier;
 import java.util.stream.IntStream;
